Preserve joint covariance precision
Keep the historical float32 covariance blocks on the legacy per-IMT path, but allow the shared covariance builder to emit float64 arrays for dense joint station, target-station, and target systems.

This avoids discarding model coefficient precision before the pseudoinverse and Schur-complement operations that define the joint reference posterior.

+++ openquake/hazardlib/tests/calc/conditioned_gmfs_test.py
@@ -168,7 +168,7 @@ def test_joint_station_covariance_combines_all_residual_components():
     phi = mean_stds_D[3, 0].reshape(-1)
     expected = compute_within_event_covariance_matrix(
         inp.within_event_model, None, distances,
-        imts, imts, phi, phi)
+        imts, imts, phi, phi, dtype=numpy.float64)
     expected = expected.astype(numpy.float64)
     numpy.fill_diagonal(
         expected, numpy.diag(expected) +
